+++ app/src/main/java/com/vitaliimalone/simpletodo/domain/usecases/GetTasksForHomeTabUseCase.kt
@@ -3,7 +3,7 @@ package com.vitaliimalone.simpletodo.domain.usecases
 import com.vitaliimalone.simpletodo.data.repository.TaskRepository
 import com.vitaliimalone.simpletodo.domain.mappers.TaskMapper
 import com.vitaliimalone.simpletodo.domain.models.Task
-import com.vitaliimalone.simpletodo.presentation.models.HomeTab
+import com.vitaliimalone.simpletodo.presentation.home.common.HomeTab
 import com.vitaliimalone.simpletodo.presentation.utils.DateTimeUtils
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
