+++ app/src/main/java/com/vitaliimalone/simpletodo/presentation/home/common/HomeTab.kt
@@ -1,4 +1,4 @@
-package com.vitaliimalone.simpletodo.presentation.models
+package com.vitaliimalone.simpletodo.presentation.home.common
 
 import com.vitaliimalone.simpletodo.R
 import com.vitaliimalone.simpletodo.presentation.utils.Res
