+++ app/src/main/java/com/vitaliimalone/simpletodo/presentation/home/common/HomeTabFragmentAdapter.kt
@@ -3,7 +3,6 @@ package com.vitaliimalone.simpletodo.presentation.home.common
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.vitaliimalone.simpletodo.presentation.hometab.HomeTabFragment
-import com.vitaliimalone.simpletodo.presentation.models.HomeTab
 
 class HomeTabFragmentAdapter(
     private val fragments: Array<HomeTab>, fragment: Fragment
