+++ app/src/main/java/com/vitaliimalone/simpletodo/presentation/settings/common/LanguageAdapter.kt
@@ -5,7 +5,6 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.vitaliimalone.simpletodo.R
-import com.vitaliimalone.simpletodo.presentation.models.Language
 import kotlinx.android.extensions.LayoutContainer
 import kotlinx.android.synthetic.main.language_dialog_item.*
