+++ app/src/main/java/com/vitaliimalone/simpletodo/presentation/utils/LanguageUtils.kt
@@ -2,7 +2,7 @@ package com.vitaliimalone.simpletodo.presentation.utils
 
 import android.content.Context
 import android.content.res.Configuration
-import com.vitaliimalone.simpletodo.presentation.models.Language
+import com.vitaliimalone.simpletodo.presentation.settings.common.Language
 import java.util.Locale
 
 // https://medium.com/ironsource-tech-blog/conversion-by-translation-changing-your-android-app-language-at-runtime-5c9daebf9771
